Add option for deny watching of subfiles for certain filesystem types
In many cases (depending on inotify_watch_add() options) libinotify-kqueue
opens all files in watched directory that can be inappropriate on certain
filesystems like network or mtpfs due to large overhead of open() operation.
That is why skip-subfiles configure option is introduced. If this option
was enabled at configure time, libinotify-kqueue checks filesystem type on
inotify_watch_add() invocation and compares with blacklist to deceide
if it necessary to watch for subfiles or not.
That disables notifications of subfiles content or attribute modifications
but still allow to track their creation, deletion and renaming.

E.g: to prohibit of opening subfiles on NFS and MTP-fs(mounted with fuse) use

./configure --enable-skip-subfiles=fusefs,nfs