Do writes of inotify events to socket buffer only if it is emtpy
Modern FreeBSDs always report full sendbuffer size in data field of
EVFILT_WRITE kevent so we can not determine amount of data remaining in
it reliably. Therefore it is still possible to check empty sendbuffer
conditions with setting of low watermark to be equal sendbuffer size
So detect that client has read all events and we can fill sendbuf again
as we want to know exact amount of bytes to avoid partial inotify event
reads as much as possible.